requirements.txt
django-toolbelt==0.0.1
django-extensions==1.2.5
django-redis>=3.3
django-rq>=0.6

Django==1.6.5
djangorestframework>=2.3
django-cors-headers==0.12

redis>=2.10
rq>=0.4
rq-scheduler>=0.4

requests>=2.3